Accept any Integer for type id in parse_abi_info
The `::Int64` typeassert at the JSON-id extraction site silently rejected
any caller whose dict carries an Int32 id, including 32-bit Julia (where
unannotated integer literals are Int32) and any future JSON backend that
returns narrower integer types. Widen to `::Integer` and let the `Int(...)`
wrapper handle the platform-Int narrowing.

The accompanying malformed-input test now explicitly constructs an Int32
id so the assertion-laxness is exercised on every platform, not only on
x86 CI.

+++ src/abi_import.jl
@@ -230,7 +230,7 @@ function parse_abi_info(parsed::AbstractDict)
     # Extract all the type descriptors
     typedescs = OrderedDict{Int, TypeDesc}()
     for type in parsed["types"]
-        id = Int(type["id"]::Int64)
+        id = Int(type["id"]::Integer)
         typedescs[id] = from_json(TypeDesc, type)
     end
